refactor: group reader options and metrics
