+++ src/lib.rs
@@ -1,7 +1,6 @@
 #![doc = include_str!("../README.md")]
 #![cfg_attr(test, allow(clippy::panic))]
 
-mod config;
 #[cfg(feature = "datafusion")]
 mod datafusion_dynamic_filters;
 #[cfg(feature = "datafusion")]
@@ -13,13 +12,11 @@ mod datafusion_planning;
 #[cfg(feature = "datafusion")]
 mod datafusion_provider;
 mod deletion_vector;
-mod direct;
 mod error;
 mod kernel;
 #[cfg(feature = "native-async")]
 #[allow(dead_code)]
 mod metered_object_store;
-mod metrics;
 #[cfg(feature = "native-async")]
 #[allow(dead_code)]
 mod native_async_reader;
@@ -33,15 +30,13 @@ mod partition_target;
 mod planning;
 mod predicate;
 mod protocol;
+mod reader;
 #[allow(dead_code)]
 mod scheduling;
 mod snapshot;
 mod transform;
 mod uri;
 
-pub use config::{
-    DeltaReaderBackend, DeltaReaderExecutionOptions, DeltaSnapshotSelection, DeltaStorageOptions,
-};
 #[cfg(feature = "datafusion")]
 pub use datafusion_execution::{
     DeltaDataFusionMetrics, DeltaDataFusionMetricsSnapshot, DeltaFileRepartitioning,
@@ -51,12 +46,7 @@ pub use datafusion_execution::{
 pub use datafusion_provider::{
     DeltaDataFusionScanOptions, DeltaTableProvider, RegisteredDeltaTable, register_delta_table,
 };
-pub use direct::{
-    DeltaBatchStream, DeltaScan, DeltaScanBuilder, DeltaTable, DeltaTableBuilder,
-    DeltaTableSnapshot,
-};
 pub use error::{DeltaReaderError, DeltaReaderPhase};
-pub use metrics::{DeltaReadMetrics, DeltaReadMetricsSnapshot};
 #[doc(hidden)]
 pub use partition_target::{
     DeltaScanPartitionTargetDiagnosticInput, DeltaScanPartitionTargetDiagnosticOutput,
@@ -67,6 +57,11 @@ pub use partition_target::{
 };
 pub use predicate::{DeltaComparison, DeltaPredicate, DeltaScalar};
 pub use protocol::DeltaProtocolInfo;
+pub use reader::{
+    DeltaBatchStream, DeltaReadMetrics, DeltaReadMetricsSnapshot, DeltaReaderBackend,
+    DeltaReaderExecutionOptions, DeltaScan, DeltaScanBuilder, DeltaSnapshotSelection,
+    DeltaStorageOptions, DeltaTable, DeltaTableBuilder, DeltaTableSnapshot,
+};
 
 /// The crate version.
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
